+++ android_app/www/js/user.service.js
@@ -1,4 +1,5 @@
-var User = function() {};
+var User = function () {
+};
 /**
  *
  * persistUserInfo
@@ -7,32 +8,39 @@ var User = function() {};
  * @param password
  * @param cb
  */
-User.prototype.saveUser = function (userId, username, password, cb){
+User.prototype.saveUser = function (userId, username, password, cb) {
   "use strict";
   var db = openDatabase('deltaplus', '1.0', 'deltaplus', 5 * 1024 * 1024);
   db.transaction(function (tx) {
     tx.executeSql('CREATE TABLE IF NOT EXISTS userInfo(userId, username, password)');
     tx.executeSql('SELECT * FROM userInfo WHERE username = ?', [username], function (tx, re) {
       if (re.rows.length == 0) {
-        tx.executeSql('INSERT INTO userInfo VALUES(?, ?, ?)', [userId, username, password], function(tx, re) {
+        tx.executeSql('INSERT INTO userInfo VALUES(?, ?, ?)', [userId, username, password], function (tx, re) {
           console.log('save user info : ' + userId)
+          cb(null, username);
+        }, function (transaction, error) {
+          console.log('save user info with error: ' + JSON.stringify(error))
+          cb(error);
         });
       } else {
-        tx.executeSql('UPDATE userInfo SET password =? WHERE username = ?', [password, username], function(tx, re) {});
+        tx.executeSql('UPDATE userInfo SET password =? WHERE username = ?', [password, username], function (tx, re) {
+        }, function (transaction, error) {
+          console.log('updated user info with error: ' + JSON.stringify(error))
+          cb(error);
+        });
       }
     });
-    cb();
   });
 };
 
-User.prototype.readUser = function (cb){
+User.prototype.readUser = function (cb) {
   "use strict";
   var db = openDatabase('deltaplus', '1.0', 'deltaplus', 5 * 1024 * 1024);
   db.transaction(function (tx) {
     tx.executeSql('CREATE TABLE IF NOT EXISTS userInfo(userId, username, password)');
-    tx.executeSql('SELECT userId, username, password FROM userInfo', null, function (tx, re) {
+    tx.executeSql('SELECT userId, username, password FROM userInfo', [], function (tx, re) {
       if (re.rows.length > 0) {
-        cb(re.rows[0]);
+        cb(re.rows.item(0));
       } else {
         cb(null)
       }
@@ -40,20 +48,23 @@ User.prototype.readUser = function (cb){
   });
 };
 
-User.prototype.clearUser = function (cb){
+User.prototype.clearUser = function (cb) {
   "use strict";
   var db = openDatabase('deltaplus', '1.0', 'deltaplus', 5 * 1024 * 1024);
   db.transaction(function (tx) {
     tx.executeSql('CREATE TABLE IF NOT EXISTS userInfo(userId, username, password)');
     tx.executeSql('DELETE FROM userInfo ', [], function (tx, re) {
       cb();
-    }), function(err) {
+    }), function (err) {
       console.log('error: ' + JSON.stringify(err))
     };
   });
 };
 
-User.prototype.autoLogin = function (callback){
+User.prototype.autoLogin = function (callback) {
+  if (!checkConnection()) {
+    return callback(new Error('No network.'));
+  }
   var loader = showLoading('Sign in ...');
   var username = window.localStorage['username'];
   var password = window.localStorage['password'];