+++ docs/ChangeLog.md
@@ -1,3 +1,13 @@
+## Changes on 2017-03-22  v1.7.2
+- Use the login page as the default page when open a app.
+- Auto sync data between app and website when user login. 2 synchronization job as below
+    - Auto upload data into website from APP DB
+    - Then auto download data into APP DB from website
+- Keep user login status even user close the app, unless user click the sign-out button.
+- When user reopen app,
+    - automatically sign-in if with network, and do data synchronization of step 2
+    - Automatically log-off user if no network.
+
 ## Changes on 2017-02-02  v1.7.1
 - Change web access link to new URL and add new link for "About Manager"
 - Rename upload/ download my product list button label